Updated main.dart. Renaming dart files to follow recommended name rules. Created create_page.dart, view_exam_page.dart, setting_page.dart. dashboard_page.dart is 90% complete.
