--- TeamB/lib/helpPage.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Help extends StatelessWidget {
-  const Help({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
