--- TeamB/lib/notificationsPage.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Notifications extends StatelessWidget {
-  const Notifications({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
