--- TeamB/lib/searchPage.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Search extends StatelessWidget {
-  const Search({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Search Page'),
-      ),
-      body: const Center(
-        child: Text('This is the search page'),
-      ),
-    );
-  }
-}
